Moved createKubeconfig function to kubectlutility
