+++ Tasks/Common/utility-common/kubectlutility.ts
@@ -7,6 +7,7 @@ import * as util from "util";
 const uuidV4 = require('uuid/v4');
 const kubectlToolName = "kubectl"
 export const stableKubectlVersion = "v1.8.9"
+var Base64 = require('js-base64').Base64;
 
 
 var fs = require('fs');
@@ -49,6 +50,19 @@ export async function downloadKubectl(version: string) : Promise<string> {
     return kubectlPath;
 }
 
+export function createKubeconfig(kubernetesServiceEndpoint: string): string
+{
+    var kubeconfigTemplateString = '{"apiVersion":"v1","kind":"Config","clusters":[{"cluster":{"certificate-authority-data": null,"server": null}}], "users":[{"user":{"token": null}}]}';
+    var kubeconfigTemplate = JSON.parse(kubeconfigTemplateString);
+
+    //populate server url, ca cert and token fields
+    kubeconfigTemplate.clusters[0].cluster.server = tl.getEndpointUrl(kubernetesServiceEndpoint, false);
+    kubeconfigTemplate.clusters[0].cluster["certificate-authority-data"] = tl.getEndpointAuthorizationParameter(kubernetesServiceEndpoint, 'serviceAccountCertificate', false);
+    kubeconfigTemplate.users[0].user.token = Base64.decode(tl.getEndpointAuthorizationParameter(kubernetesServiceEndpoint, 'serviceAccountToken', false));
+
+    return JSON.stringify(kubeconfigTemplate);
+}
+
 function getTempDirectory(): string {
     return tl.getVariable('agent.tempDirectory') || os.tmpdir();
 }
@@ -76,4 +90,3 @@ function getExecutableExtention(): string {
 
     return "";
 }
-
