+++ Tasks/Common/utility-common/package.json
@@ -16,6 +16,7 @@
     "semver": "^5.4.1",
     "vso-node-api": "6.0.1-preview",
     "vsts-task-lib": "2.0.6",
-    "vsts-task-tool-lib": "0.4.0"
+    "vsts-task-tool-lib": "0.4.0",
+    "js-base64": "2.4.3"
   }
 }
