+++ Tasks/KubernetesV0/make.json
@@ -15,7 +15,9 @@
         {
             "items": [
                 "node_modules/utility-common/node_modules/vsts-task-lib",
-                "node_modules/docker-common/node_modules/vsts-task-lib"
+                "node_modules/docker-common/node_modules/vsts-task-lib",
+                "node_modules/vsts-task-tool-lib/node_modules/vsts-task-lib",
+                "node_modules/azure-arm-rest/node_modules/vsts-task-lib"
             ],
             "options": "-Rf"
         }
