+++ Tasks/KubernetesV0/package.json
@@ -2,7 +2,6 @@
   "dependencies": {
     "del": "2.2.0",
     "docker-common": "file:../../_build/Tasks/Common/docker-common-1.0.0.tgz",
-    "js-base64": "2.4.3",
     "utility-common": "file:../../_build/Tasks/Common/utility-common-1.0.2.tgz",
     "vsts-task-tool-lib": "0.4.0"
   }
